refactor(sessions): 拆分会话存储与保留策略子模块
