+++ pinvou3-app/src-tauri/src/features/sessions/validators.rs
@@ -1,0 +1,119 @@
+//! Session id, workspace, and path validators plus the small persistence
+//! helpers shared across the session store submodules.
+//!
+//! These free functions are intentionally side-effect-free (validation) or
+//! trivially derivable (path composition / system-prompt flattening) so that
+//! every other submodule can depend on them without introducing cycles.
+
+use std::path::{Component, Path, PathBuf};
+
+use anyhow::{bail, Result};
+use deepseek_tui::models::SystemPrompt;
+use deepseek_tui::session_manager::SessionManager;
+
+/// 生成 URL-safe session id（短 8 字节 timestamp + nanos hash）。
+/// 上游 `validated_session_path` 只允许 `[A-Za-z0-9_-]`，所以走 base32-like 字符集。
+
+impl SessionStore {
+pub(crate) fn validate_session_id(id: &str) -> Result<()> {
+    if id.trim().is_empty()
+        || !id.chars().all(|character| {
+            character.is_ascii_alphanumeric() || character == '-' || character == '_'
+        })
+    {
+        bail!("Invalid session id '{id}'");
+    }
+    Ok(())
+}
+
+fn validate_scheduled_session_id(id: &str) -> Result<()> {
+    validate_session_id(id)?;
+    if !id.starts_with("sched-") {
+        bail!("Scheduled session id must start with 'sched-': {id}");
+    }
+    Ok(())
+}
+
+fn validate_scheduled_task_id(id: &str) -> Result<()> {
+    if id.trim().is_empty()
+        || !id.chars().all(|character| {
+            character.is_ascii_alphanumeric() || character == '-' || character == '_'
+        })
+    {
+        bail!("Invalid scheduled task id '{id}'");
+    }
+    Ok(())
+}
+
+fn validate_scheduled_workspace_path(root: &Path, workspace: &Path) -> Result<()> {
+    if !workspace.is_absolute() {
+        bail!(
+            "Scheduled profile workspace must be absolute: {}",
+            workspace.display()
+        );
+    }
+    if workspace
+        .components()
+        .any(|component| matches!(component, Component::ParentDir))
+    {
+        bail!(
+            "Scheduled profile workspace must not contain parent segments: {}",
+            workspace.display()
+        );
+    }
+    if !workspace.starts_with(root) {
+        bail!(
+            "Scheduled profile workspace must live under {}: {}",
+            root.display(),
+            workspace.display()
+        );
+    }
+    if workspace.file_name().and_then(|name| name.to_str()) != Some("workspace") {
+        bail!(
+            "Scheduled profile workspace must end with 'workspace': {}",
+            workspace.display()
+        );
+    }
+    Ok(())
+}
+
+fn persisted_system_prompt(system_prompt: Option<&SystemPrompt>) -> Option<String> {
+    match system_prompt {
+        Some(SystemPrompt::Text(text)) => Some(text.clone()),
+        Some(SystemPrompt::Blocks(blocks)) => Some(
+            blocks
+                .iter()
+                .map(|block| block.text.clone())
+                .collect::<Vec<_>>()
+                .join("\n\n---\n\n"),
+        ),
+        None => None,
+    }
+}
+
+fn chat_session_file(manager: &SessionManager, id: &str) -> Result<PathBuf> {
+    validate_session_id(id)?;
+    Ok(manager.sessions_dir().join(format!("{id}.json")))
+}
+
+fn scheduled_session_file(manager: &SessionManager, id: &str) -> Result<PathBuf> {
+    validate_scheduled_session_id(id)?;
+    Ok(manager.sessions_dir().join(format!("{id}.json")))
+}
+
+fn generate_session_id() -> String {
+    use std::time::{SystemTime, UNIX_EPOCH};
+    let nanos = SystemTime::now()
+        .duration_since(UNIX_EPOCH)
+        .map(|d| d.as_nanos())
+        .unwrap_or(0);
+    const ALPHA: &[u8] = b"0123456789abcdefghijklmnopqrstuvwxyz";
+    let mut n = nanos;
+    let mut buf = String::with_capacity(13);
+    for _ in 0..13 {
+        buf.push(ALPHA[(n % 36) as usize] as char);
+        n /= 36;
+    }
+    buf
+}
+}
